se agrego sweetalert2.js

+++ src/components/UsersList.jsx
@@ -1,24 +1,60 @@
 import React from 'react'
 import axios from 'axios'
+import Swal from "sweetalert2";
+import 'animate.css';
 
-const UsersList = ({user,allUsers,setUpdateData,handleOpenForm,handleMesajeDelete,handleCloseForm}) => {
+const UsersList = ({user,allUsers,setUpdateData,handleOpenForm}) => {
     // console.log(user);
 
  const deteleUsers = () => {
-     const URL = `https://users-crud1.herokuapp.com/users/${user.id}/`
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!',
+    showClass: {
+      popup: 'animate__animated animate__fadeInDown'
+    },
+    hideClass: {
+      popup: 'animate__animated animate__fadeOutUp'
+    }
+  }).then((res) => {
+    if (res.isConfirmed) {  
+      Swal.fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success',
+      )
+      const URL = `https://users-crud1.herokuapp.com/users/${user.id}/`
       axios.delete(URL)
-      .then(res => {
-        // console.log(res.data);
-        handleCloseForm()
-        handleMesajeDelete()
-        allUsers()
-      })
+      allUsers()
+    }
+  })
+  
+
+
+
+
+     
+      // .then(res => {
+      //   // console.log(res.data);
+      //   alertDelete()
+        
+       
+      // })
       .catch(err => console.log(err))
 }
  
 const bottonUpdate = () =>{
     setUpdateData(user)
     handleOpenForm() 
+}
+
+const alertDelete = () => {
+  deteleUsers()
 }
 
 
@@ -35,7 +71,7 @@ const bottonUpdate = () =>{
            <li>  <span>Birthday:</span> {user.birthday}</li> 
         </ul>
         <footer className='footer'>
-            <button className='button__delete' onClick={deteleUsers}><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3" viewBox="0 0 16 16">
+            <button className='button__delete' onClick={alertDelete}><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3" viewBox="0 0 16 16">
                     <path d="M6.5 1h3a.5.5 0 0 1 .5.5v1H6v-1a.5.5 0 0 1 .5-.5ZM11 2.5v-1A1.5 1.5 0 0 0 9.5 0h-3A1.5 1.5 0 0 0 5 1.5v1H2.506a.58.58 0 0 0-.01 0H1.5a.5.5 0 0 0 0 1h.538l.853 10.66A2
                      2 0 0 0 4.885 16h6.23a2 2 0 0 0 1.994-1.84l.853-10.66h.538a.5.5 0 0 0 0-1h-.995a.59.59 0 0 0-.01 0H11Zm1.958 1-.846 10.58a1 1 0 0 1-.997.92h-6.23a1 1 0 0 1-.997-.92L3.042
                       3.5h9.916Zm-7.487 1a.5.5 0 0 1 .528.47l.5 8.5a.5.5 0 0 1-.998.06L5 5.03a.5.5 0 0 1 .47-.53Zm5.058 0a.5.5 0 0 1 .47.53l-.5 8.5a.5.5 0 1 1-.998-.06l.5-8.5a.5.5 0 0 1 .528-.47ZM8 4.5a.5.5 0 0 1 .5.5v8.5a.5.5 0 0 1-1 0V5a.5.5 0 0 1 .5-.5Z"/>
